Game Logic Finalized
Now, one of the destination points is randomly selected to become the player's objective. Additionally, at the beginning of the game, the truck is loaded with 10 crates. When the player reaches the destination with the vehicle, one load is considered delivered. Once all the loads are completed, the truck is reloaded with crates again.