Fuel System Added
As time progresses, the fuel decreases. To refill the fuel, the player needs to go to the gas station and collect the fuel can item. When the fuel is depleted, the car does not take gas, and if the player cannot reach the fuel can during this time period, they lose because the fuel has been depleted.