Nitrous System Added
The player increases the acceleration of the vehicle by holding down the left shift key. During the time they hold it down, the nitrous level decreases, and once it is completely depleted, they can no longer use nitro. To acquire more nitro, the player needs to find and touch a pickable item, a nitrous tube.